+++ apps/vscode/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.83.0 (Unreleased)
 
 - Correct syntax highlighting for list items and blockquote character
+- Enable Quarto: Render Project command for Hugo and Docusausus projects
 
 ## 1.82.0 (Release on 21 April 2023)